Adds arrays for location, speed, etc, and summary activity page
The activities model has been changed to account for many POST requests
that will add data to arrays in one Activity document. /activities/add
has been split into activities/init and activities/add, where init is
called when starting a new activity.
The activitiysummary page is loaded when an activity is selected in
recentactivities page, the date (Date obj) is sent as a url parameter,
in order to get the required data. Activity type can also be updated
here, also using the date parameter in the url.